pyproject.toml
[project]
name = "hostingserver"
version = "0.1.0"
description = "local server for hosting; dockerized"
authors = [
    {name = "Cstannahill",email = "cstannahill10@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "uvicorn (>=0.34.3,<0.35.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "pyyaml",
    "jinja2"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
